move src/dxapi_portal_runner/main.py back to run.py
